requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
PyJWT>=2.8.0
requests>=2.31.0
django-cors-headers>=4.3.0
python-decouple>=3.8

# Database
sqlalchemy==2.0.23
mysqlclient==2.2.0
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and HTTP
fastapi-cors==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 